Initial commit.
Classes are still over the place. Will have to clean up.

I think having a dedicated event store per service is the way
to go here. In this case, the Domain Event entity is solely for the use
of Inventory domain events.

Also going for storing data related to the event instead of passing
references. It's the easier approach for I think.